requirements.txt
channels==3.0.5
channels-redis==4.1.0
Django==4.1.5
django-crispy-forms==1.14.0
psycopg2==2.9.5
python-dotenv==1.0.0
redis==4.5.4
requests~=2.28.2